[FAGSYSTEM-194254] flytte reducer-kall til meldinger
Vi ønsker å gjøre kallet til navkontor-reduceren lenger ned
i komponenttreet.
