+++ src/app/personside/infotabs/meldinger/MeldingerContainer.tsx
@@ -6,7 +6,7 @@ import { pxToRem } from '../../../../styles/personOversiktTheme';
 import TraadListe from './traadliste/TraadListe';
 import { huskSokAction, setSkjulVarslerAction } from '../../../../redux/meldinger/actions';
 import { useDispatch } from 'react-redux';
-import { useAppState, usePrevious } from '../../../../utils/customHooks';
+import { useAppState, useOnMount, usePrevious } from '../../../../utils/customHooks';
 import { useInfotabsDyplenker } from '../dyplenker';
 import { useHistory, withRouter } from 'react-router';
 import { AlertStripeFeil, AlertStripeInfo } from 'nav-frontend-alertstriper';
@@ -17,6 +17,7 @@ import TraadVisningWrapper from './traadvisning/TraadVisningWrapper';
 import { useRestResource } from '../../../../rest/consumer/useRestResource';
 import DelayRender from '../../../../components/DelayRender';
 import { useKeepQueryParams } from '../../../../utils/hooks/useKeepQueryParams';
+import { isNotStarted } from '../../../../rest/utils/restResource';
 
 const meldingerMediaTreshold = pxToRem(800);
 
@@ -63,9 +64,9 @@ function useHuskSokeord(sokeord: string) {
 
 function useReloadOnEnhetChange() {
     const dispatch = useDispatch();
-    const enhet = useAppState(state => state.session.valgtEnhetId);
+    const enhet = useAppState((state) => state.session.valgtEnhetId);
     const forrigeEnhet = usePrevious(enhet);
-    const meldingerResource = useRestResource(resources => resources.traader);
+    const meldingerResource = useRestResource((resources) => resources.traader);
 
     useEffect(() => {
         if (!forrigeEnhet) {
@@ -77,20 +78,31 @@ function useReloadOnEnhetChange() {
     }, [forrigeEnhet, enhet, meldingerResource, dispatch]);
 }
 
+function useBrukersNavKontor() {
+    const resource = useAppState((state) => state.restResources.brukersNavKontor);
+    const dispatch = useDispatch();
+
+    useOnMount(() => {
+        if (isNotStarted(resource)) {
+            dispatch(resource.actions.fetch);
+        }
+    });
+}
+
 function MeldingerContainer() {
     const dispatch = useDispatch();
-    const traaderResource = useRestResource(resources => resources.traader, undefined, true);
-    const skjulVarsler = useAppState(state => state.meldinger.skjulVarsler);
+    const traaderResource = useRestResource((resources) => resources.traader, undefined, true);
+    const skjulVarsler = useAppState((state) => state.meldinger.skjulVarsler);
     const setSkjulVarsler = (skjul: boolean) => dispatch(setSkjulVarslerAction(skjul));
-    const forrigeSok = useAppState(state => state.meldinger.forrigeSok);
+    const forrigeSok = useAppState((state) => state.meldinger.forrigeSok);
     const [sokeord, setSokeord] = useState(forrigeSok);
 
     const traaderFørSøk = traaderResource.data ? traaderResource.data : [];
-    const traaderEtterSokOgFiltrering = useSokEtterMeldinger(traaderFørSøk, sokeord).filter(traad =>
+    const traaderEtterSokOgFiltrering = useSokEtterMeldinger(traaderFørSøk, sokeord).filter((traad) =>
         skjulVarsler ? filtrerBortVarsel(traad) : true
     );
     const valgtTraad = useValgtTraadIUrl() || traaderEtterSokOgFiltrering[0];
-
+    useBrukersNavKontor();
     useKeepQueryParams();
     useSyncSøkMedVisning(traaderFørSøk, traaderEtterSokOgFiltrering, valgtTraad);
     useHuskSokeord(sokeord);
